requirements.txt
asgiref==3.7.2
cryptography==39.0.1
Django==4.2.4
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.14.0
mypy==1.5.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.7
pycryptodome==3.18.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0